+++ src/main/java/com/carrental/dto/CarStatusDto.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @Getter
 public class CarStatusDto {
 
-    private int CarStatusId;
+    private long CarStatusId;
     private String CarStatusDescription;
 }
